Add multiline input support until semicolon

+++ myDB/app.h
@@ -5,7 +5,7 @@
 void runDB(AppContext * app);
 void shutdownDB(AppContext* app);
 void printCurrentDbState(AppContext * app);
-int readInput(char *buffer, size_t size);
+int readInput(char *buffer, int size);
 int checkExit(AppContext* app, char** argv);
 void handleCommand(AppContext* app, char** argv, int argc);
